chore(api): move /possibilities route  into prescription context
