--- api/lib/application/sco-organization-learners/index.js
@@ -1,45 +1,0 @@
-import JoiDate from '@joi/date';
-import BaseJoi from 'joi';
-const Joi = BaseJoi.extend(JoiDate);
-
-import { scoOrganizationLearnerController } from '../../../src/prescription/organization-learner/application/sco-organization-learner-controller.js';
-import { sendJsonApiError, UnprocessableEntityError } from '../../../src/shared/application/http-errors.js';
-
-const register = async function (server) {
-  server.route([
-    {
-      method: 'PUT',
-      path: '/api/sco-organization-learners/possibilities',
-      config: {
-        auth: false,
-        handler: scoOrganizationLearnerController.generateUsername,
-        validate: {
-          options: {
-            allowUnknown: true,
-          },
-          payload: Joi.object({
-            data: {
-              attributes: {
-                'first-name': Joi.string().trim().empty(null).required(),
-                'last-name': Joi.string().trim().empty(null).required(),
-                birthdate: Joi.date().format('YYYY-MM-DD').raw().required(),
-                'organization-id': Joi.number().empty(null).required(),
-              },
-            },
-          }),
-          failAction: (request, h) => {
-            return sendJsonApiError(new UnprocessableEntityError('Un des champs saisis n’est pas valide.'), h);
-          },
-        },
-        notes: [
-          '- Elle permet de savoir si un élève identifié par son nom, prénom et date de naissance est inscrit à ' +
-            "l'organisation détenant la campagne. Cet élève n'est, de plus, pas encore associé à l'organisation.",
-        ],
-        tags: ['api', 'sco-organization-learners'],
-      },
-    },
-  ]);
-};
-
-const name = 'sco-organization-learners-api';
-export { name, register };
