--- api/lib/routes.js
@@ -1,5 +1,0 @@
-import * as scoOrganizationLearners from './application/sco-organization-learners/index.js';
-
-const routes = [scoOrganizationLearners];
-
-export { routes };
