--- api/tests/integration/application/sco-organization-learners/index_test.js
@@ -1,181 +1,0 @@
-import * as moduleUnderTest from '../../../../lib/application/sco-organization-learners/index.js';
-import { libScoOrganizationLearnerController } from '../../../../lib/application/sco-organization-learners/sco-organization-learner-controller.js';
-import { securityPreHandlers } from '../../../../src/shared/application/security-pre-handlers.js';
-import { expect, HttpTestServer, sinon } from '../../../test-helper.js';
-
-describe('Integration | Application | Route | sco-organization-learners', function () {
-  let httpTestServer;
-
-  beforeEach(async function () {
-    sinon
-      .stub(libScoOrganizationLearnerController, 'generateUsername')
-      .callsFake((request, h) => h.response('ok').code(200));
-    sinon
-      .stub(securityPreHandlers, 'checkUserBelongsToScoOrganizationAndManagesStudents')
-      .callsFake((request, h) => h.response(true));
-
-    httpTestServer = new HttpTestServer();
-    await httpTestServer.register(moduleUnderTest);
-  });
-
-  describe('PUT /api/sco-organization-learners/possibilities', function () {
-    const method = 'PUT';
-    const url = '/api/sco-organization-learners/possibilities';
-
-    it('should exist', async function () {
-      // given
-      const payload = {
-        data: {
-          attributes: {
-            'first-name': 'Robert',
-            'last-name': 'Smith',
-            birthdate: '2012-12-12',
-            'organization-id': 1,
-          },
-        },
-      };
-
-      // when
-      const response = await httpTestServer.request(method, url, payload);
-
-      // then
-      expect(response.statusCode).to.equal(200);
-    });
-
-    it('should return an error when there is no payload', async function () {
-      // when
-      const response = await httpTestServer.request(method, url);
-
-      // then
-      expect(response.statusCode).to.equal(422);
-    });
-
-    it('should return an error when there is an invalid first name attribute in the payload', async function () {
-      // given
-      const INVALID_FIRSTNAME = ' ';
-      const payload = {
-        data: {
-          attributes: {
-            'first-name': INVALID_FIRSTNAME,
-            'last-name': 'Smith',
-            birthdate: '2012-12-12',
-            'organization-id': 1,
-          },
-        },
-      };
-
-      // when
-      const response = await httpTestServer.request(method, url, payload);
-
-      // then
-      expect(response.statusCode).to.equal(422);
-    });
-
-    it('should return an error when there is an invalid last name attribute in the payload', async function () {
-      // given
-      const INVALID_LASTNAME = '';
-      const payload = {
-        data: {
-          attributes: {
-            'first-name': 'Robert',
-            'last-name': INVALID_LASTNAME,
-            birthdate: '2012-12-12',
-            'organization-id': 1,
-          },
-        },
-      };
-
-      // when
-      const response = await httpTestServer.request(method, url, payload);
-
-      // then
-      expect(response.statusCode).to.equal(422);
-    });
-
-    it('should return an error when there is an invalid a birthdate attribute (with space) in the payload', async function () {
-      // given
-      const INVALID_BIRTHDATE = '2012- 12-12';
-      const payload = {
-        data: {
-          attributes: {
-            'first-name': 'Robert',
-            'last-name': 'Smith',
-            birthdate: INVALID_BIRTHDATE,
-            'organization-id': 1,
-          },
-        },
-      };
-
-      // when
-      const response = await httpTestServer.request(method, url, payload);
-
-      // then
-      expect(response.statusCode).to.equal(422);
-    });
-
-    it('should return an error when there is an invalid birthdate attribute (with extra zeros) in the payload', async function () {
-      // given
-      const INVALID_BIRTHDATE = '2012-012-12';
-
-      // when
-      const payload = {
-        data: {
-          attributes: {
-            'first-name': 'Robert',
-            'last-name': 'Smith',
-            birthdate: INVALID_BIRTHDATE,
-            'organization-id': 1,
-          },
-        },
-      };
-
-      // when
-      const response = await httpTestServer.request(method, url, payload);
-
-      // then
-      expect(response.statusCode).to.equal(422);
-    });
-
-    it('should return an error when there is an invalid birthdate attribute (not a proper date) in the payload', async function () {
-      // given
-      const INVALID_BIRTHDATE = '1999-99-99';
-      const payload = {
-        data: {
-          attributes: {
-            'first-name': 'Robert',
-            'last-name': 'Smith',
-            birthdate: INVALID_BIRTHDATE,
-            'organization-id': 1,
-          },
-        },
-      };
-
-      // when
-      const response = await httpTestServer.request(method, url, payload);
-
-      // then
-      expect(response.statusCode).to.equal(422);
-    });
-
-    it('should return an error when there is an invalid campaign code attribute in the payload', async function () {
-      // given
-      const INVALID_CAMPAIGNCODE = '';
-      const payload = {
-        data: {
-          attributes: {
-            'first-name': 'Robert',
-            'last-name': 'Smith',
-            birthdate: '2012-12-12',
-            'organization-id': INVALID_CAMPAIGNCODE,
-          },
-        },
-      };
-
-      // when
-      const response = await httpTestServer.request(method, url, payload);
-
-      // then
-      expect(response.statusCode).to.equal(422);
-    });
-  });
-});
